requirements.txt
openai-whisper
sounddevice
numpy
keyboard
pyperclip
wavio
torch
plyer
ollama
